chore(refactor): Move env notification handling to new file. Refactor it
